+++ felupe/element/__init__.py
@@ -6,6 +6,6 @@ from ._hexahedron import (
 )
 from ._lagrange import ArbitraryOrderLagrange
 from ._line import Line
-from ._quad import ConstantQuad, Quad
+from ._quad import BiQuadraticQuad, ConstantQuad, Quad
 from ._tetra import QuadraticTetra, Tetra, TetraMINI
 from ._triangle import QuadraticTriangle, Triangle, TriangleMINI